+++ static/app/views/detectors/components/detectorLink.tsx
@@ -77,6 +77,11 @@ function DetailItem({children}: {children: React.ReactNode}) {
 }
 
 function ConfigDetails({detector}: {detector: Detector}) {
+  // TODO: Use a MetricDetector type to avoid checking for this
+  if (!('detection_type' in detector.config)) {
+    return null;
+  }
+
   const type = detector.config.detection_type;
   const conditions = detector.conditionGroup?.conditions;
   if (!conditions?.length) {
