+++ static/app/views/detectors/constants.tsx
@@ -1,7 +1,15 @@
+import {t} from 'sentry/locale';
 import type {DetectorType} from 'sentry/types/workflowEngine/detectors';
 import {FieldValueType} from 'sentry/utils/fields';
 
 export const DETECTOR_LIST_PAGE_LIMIT = 20;
+
+export const DETECTOR_TYPE_LABELS: Record<DetectorType, string> = {
+  metric_issue: t('Metric'),
+  uptime_domain_failure: t('Uptime'),
+  error: t('Error'),
+  uptime_subscription: t('Uptime'),
+};
 
 export const DETECTOR_FILTER_KEYS: Record<
   string,
